requirements.txt
annoy==1.17.0
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
MarkupSafe==1.1.1
numpy==1.19.2
pandas==1.1.2
plotly==4.10.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
retrying==1.3.3
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
SQLAlchemy==1.3.19
threadpoolctl==2.1.0
tqdm==4.49.0
Werkzeug==1.0.1
